==> 1.Reddit_API_vF.ipynb <==
import numpy as np
import pandas as pd
import os
import requests

## Validate Credentials

# Credentials for Reddit PRAW API (redacted)

import praw
r = praw.Reddit(client_id='YOUR_ID',
                     client_secret='YOUR_SECRET',
                     password='YOUR_PASSWORD',
                     user_agent='YOUR_AGENT',
                     username='YOUR_USERNAME')

print(r.user.me())

## API Calls for Submissions and Comments

# Pulls submissions and URLs and exports into CSV
subreddits = ['SandersForPresident','socialism','Anarchism','democrats','Progressive','communism','Liberal','LateStageCapitalism',
              'The_Donald','Libertarian','Conservative','Republican','Anarcho_Capitalism',
             'politics','NeutralPolitics']

# Maximum number of submissions set at 1,000
sublimit = 1000

for i in subreddits:
    ids = []
    sub = []
    url = []
    s_score = []

    for submission in r.subreddit(i).top(limit=sublimit):
        ids.append(submission.id)
        sub.append(submission.title)
        url.append(submission.url)
        s_score.append(submission.score)

    print('Subreddit:',i)
    print('Number of Submissions:',len(sub))

    # Export submissions
    submissions = pd.DataFrame({'id':ids,'title':sub,'url':url,'score':s_score})
    sub_title = i+'_submissions.csv'
    submissions.to_csv(sub_title,index=False)


# Comments

# Pulls comments and exports into CSV

# Number of top submissions (subject to 'more comment' limit specified below)
comlimit = 1000

for i in subreddits:
    author = []
    comm = []
    c_score = []

    for submission in r.subreddit(i).top(limit=comlimit):
        submission.comments.replace_more(limit=0)
        for top_level_comment in submission.comments:
            author.append(top_level_comment.author)
            comm.append(top_level_comment.body)
            c_score.append(top_level_comment.score)

    print('Number of Comments from {} Submissions: {}'.format(comlimit,len(comm)))
            
    # Export comments
    comments = pd.DataFrame({'author':author,'comment':comm,'score':c_score})
    comm_title = i+'_comments.csv'
    comments.to_csv(comm_title,index=False)
