==> 2.Reddit_Comments_vF.ipynb <==
import pandas as pd
import numpy as np

import json

# Converts JSON to JSONReader object which can be used as an iterator
chunksize = 5e5
df = pd.read_json('massive_comment_dump', lines=True,chunksize=chunksize)

# Converts object to a list of dataframes which can be later parsed / accessed
# Need to do this with only up to 5 dataframes at a time to avoid out of memory errors 
subreddit1 = 'LateStageCapitalism' 
subreddit2 = 'Libertarian'
subreddit3 = 'Anarchism'
subreddit4 = 'politics'
subreddit5 = 'Conservative'

sub_df1 = pd.DataFrame([])
sub_df2 = pd.DataFrame([])
sub_df3 = pd.DataFrame([])
sub_df4 = pd.DataFrame([])
sub_df5 = pd.DataFrame([])

for i in df:
    sub_df1 = pd.concat([sub_df1,i[i['subreddit']==subreddit1]]) 
    sub_df2 = pd.concat([sub_df2,i[i['subreddit']==subreddit2]])
    sub_df3 = pd.concat([sub_df3,i[i['subreddit']==subreddit3]])
    sub_df4 = pd.concat([sub_df4,i[i['subreddit']==subreddit4]])
    sub_df5 = pd.concat([sub_df5,i[i['subreddit']==subreddit5]])

print(len(sub_df1))
print(len(sub_df2))
print(len(sub_df3))
print(len(sub_df4))
print(len(sub_df5))

# Update subreddit of interest to manuall inspect comments and preprocess
sub_df = sub_df5.copy()
subreddit = subreddit5

sub_df.head().transpose()

# Select only columns used in analysis
output = sub_df[['author','body','score','created_utc','id','link_id','parent_id']].copy()

# Remove deleted comments
output = output[output['body']!='[deleted]']
len(output)

# Convert timestamp
conv_time = pd.to_datetime(output['created_utc'], unit='s')

output['hour'] = conv_time.dt.hour
output['day'] = conv_time.dt.day
output['month'] = conv_time.dt.month
output['year'] = conv_time.dt.year

print('Subreddit:',subreddit,'output')
print('Number of comments:',len(output))
print()

print(output.dtypes)
print()

for i in output.columns[2:4]:
    print(i,":",np.all(np.isfinite(output[i])))
print()
    
print(output.isnull().sum())

# Remove duplicate columns 
output = output.drop_duplicates(subset="body")

# Export comments in CSV
comm_title = subreddit+'_comments.csv'
output.to_csv(comm_title,index=False)

"""comm_title = subreddit+'_comments.JSON'
output.to_json(comm_title,orient='records', lines=True)"""
